pyproject.toml
[tool.poetry]
name = "autofeeder"
version = "0.1.0"
description = "A data pipeline application to generate daily feeder cattle reports"
authors = ["John Rogers <johnr.is.coding@gmail.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"
sqlalchemy = "^2.0.46"
alembic = "^1.18.4"
aiohttp = "^3.13.3"
asyncpg = "^0.31.0"
loguru = "^0.7.3"
typer = "^0.23.0"
sqlmodel = "^0.0.33"
python-dateutil = "^2.9.0.post0"
alembic-postgresql-enum = "^1.9.0"
psycopg2 = "^2.9.11"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
black = "^26.1.0"
isort = "^7.0.0"
mypy = "^1.19.1"

[tool.poetry.scripts]
test = "scripts.run_tests:main"
snippet = "scripts.run_tests:test_snippet"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: tests that require a live database and API connection",
]

[tool.mypy]
plugins = ["pydantic.mypy"]
python_version = 3.12
check_untyped_defs = true
disallow_any_generics = true
disallow_untyped_defs = true
ignore_missing_imports = true
